fix: separate styles and folders
