+++ frontend/src/assets/styles/Background/FloatingShapes.css
@@ -1,0 +1,98 @@
+.shapes-container {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    z-index: -1;
+  }
+  
+  .custom-shape {
+    position: absolute;
+    width: 100vw;
+    opacity: 0;
+    z-index: -1;
+    background-size: cover;
+    will-change: transform, opacity;
+  }
+  
+  .top-right {
+    right: -27vw;
+    top: -22vw;
+    animation:
+      fadeIn 1s ease-in forwards,
+      float 30s infinite ease-in-out;
+  }
+  
+  .bottom-left {
+    left: -27vw;
+    bottom: -30vw;
+    animation:
+      fadeIn 1s ease-in forwards,
+      float-mirrored 30s infinite ease-in-out;
+  }
+  
+  @keyframes float {
+    0% {
+      transform: translate(0, 0) rotate(0deg);
+      opacity: 0.7;
+    }
+    50% {
+      transform: translate(-7vw, -1vw) rotate(5deg);
+      opacity: 1;
+    }
+    100% {
+      transform: translate(0, 0) rotate(0deg);
+      opacity: 0.7;
+    }
+  }
+  
+  @keyframes float-mirrored {
+    0% {
+      transform: translate(0, 0) rotate(0deg);
+      opacity: 1;
+    }
+    50% {
+      transform: translate(7vw, -1vw) rotate(15deg);
+      opacity: 0.7;
+    }
+    100% {
+      transform: translate(0, 0) rotate(0deg);
+      opacity: 1;
+    }
+  }
+  
+  @keyframes fadeIn {
+    0% {
+      opacity: 0;
+    }
+    50% {
+      opacity: 0.5;
+    }
+    100% {
+      opacity: 0.7;
+    }
+  }
+  
+  @media (max-width: 1000px) {
+  
+    .custom-shape {
+      width: 90vw;
+    }
+  
+    .top-right {
+      right: -15vw;
+      top: -15vw;
+      animation:
+        fadeIn 1s ease-in forwards,
+        float 30s infinite ease-in-out;
+    }
+    
+    .bottom-left {
+      left: -27vw;
+      bottom: -30vw;
+      animation:
+        fadeIn 1s ease-in forwards,
+        float-mirrored 30s infinite ease-in-out;
+    }
+   }
